// ya-ryadom-ui/src/utils/enums/panels.enum.ts
export enum PANELS {
    USER_EVENTS_PANEL = 'USER_EVENTS_PANEL',
    USER_REVIEWS_PANEL = 'USER_REVIEWS_PANEL',
    MY_PROFILE_PANEL = 'MY_PROFILE_PANEL',
    PROFILE_PANEL = 'PROFILE_PANEL',
    MY_PROFILE_EDIT_PANEL = 'MY_PROFILE_EDIT_PANEL',
    MY_PROFILE_EDIT_THEMES_PANEL = 'MY_PROFILE_EDIT_THEMES_PANEL',
    CREATE_EVENT_PANEL = 'CREATE_EVENT_PANEL',
    EVENTS_NEAR_ME_PANEL = 'EVENTS_NEAR_ME_PANEL',
    SELECTED_EVENT_PANEL = 'SELECTED_EVENT_PANEL',
    HELLO_INTRO_PANEL = 'HELLO_INTRO_PANEL',
    GEOLOCATION_INTRO_PANEL = 'GEOLOCATION_INTRO_PANEL',
    THEMES_INTRO_PANEL = 'THEMES_INTRO_PANEL',
    SELECT_CITY_INTRO_PANEL = 'SELECT_CITY_INTRO_PANEL',
    ABOUT_MYSELF_INTRO_PANEL = 'ABOUT_MYSELF_INTRO_PANEL',
    MY_PROFILE_EDIT_ABOUT_MYSELF_PANEL = 'MY_PROFILE_ABOUT_MYSELF_PANEL',
    COMPLETED_INTRO_PANEL = 'COMPLETED_INTRO_PANEL',
    APPLICATIONS_PANEL = 'APPLICATIONS_PANEL',
    REVIEWS_PANEL = 'REVIEWS_PANEL',
}